% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_functions.R
\name{groups.removeMember}
\alias{groups.removeMember}
\title{Removes users from the specified group.}
\usage{
groups.removeMember(GroupsRemoveMemberRequest, project, groupName)
}
\arguments{
\item{GroupsRemoveMemberRequest}{The \link{GroupsRemoveMemberRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{groupName}{Name of the group for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud.useraccounts
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud.useraccounts)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/compute/docs/access/user-accounts/api/latest/}{Google Documentation}

Other GroupsRemoveMemberRequest functions: \code{\link{GroupsRemoveMemberRequest}}
}

