% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{LinuxGroupView}
\alias{LinuxGroupView}
\title{LinuxGroupView Object}
\usage{
LinuxGroupView(gid = NULL, groupName = NULL, members = NULL)
}
\arguments{
\item{gid}{[Output Only] The Group ID}

\item{groupName}{[Output Only] Group name}

\item{members}{[Output Only] List of user accounts that belong to the group}
}
\value{
LinuxGroupView object
}
\description{
LinuxGroupView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A detailed view of a Linux group.
}

