% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{LinuxUserView}
\alias{LinuxUserView}
\title{LinuxUserView Object}
\usage{
LinuxUserView(gecos = NULL, gid = NULL, homeDirectory = NULL,
  shell = NULL, uid = NULL, username = NULL)
}
\arguments{
\item{gecos}{[Output Only] The GECOS (user information) entry for this account}

\item{gid}{[Output Only] User's default group ID}

\item{homeDirectory}{[Output Only] The path to the home directory for this account}

\item{shell}{[Output Only] The path to the login shell for this account}

\item{uid}{[Output Only] User ID}

\item{username}{[Output Only] The username of the account}
}
\value{
LinuxUserView object
}
\description{
LinuxUserView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A detailed view of a Linux user account.
}

