% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{LogConfig}
\alias{LogConfig}
\title{LogConfig Object}
\usage{
LogConfig(counter = NULL)
}
\arguments{
\item{counter}{Counter options}
}
\value{
LogConfig object
}
\description{
LogConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies what kind of log the caller must write
}
