% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{PublicKey}
\alias{PublicKey}
\title{PublicKey Object}
\usage{
PublicKey(creationTimestamp = NULL, description = NULL,
  expirationTimestamp = NULL, fingerprint = NULL, key = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{expirationTimestamp}{Optional expiration timestamp}

\item{fingerprint}{[Output Only] The fingerprint of the key is defined by RFC4716 to be the MD5 digest of the public key}

\item{key}{Public key text in SSH format, defined by RFC4253 section 6}
}
\value{
PublicKey object
}
\description{
PublicKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A public key for authenticating to guests.
}
\seealso{
Other PublicKey functions: \code{\link{users.addPublicKey}}
}

